moves selectmenublock to groups folder
