+++ src/groups/SelectMenuBlock/SelectMenuBlock.scss
